requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
aiohttp==3.8.1
    # via
    #   eth-brownie
    #   web3
aiosignal==1.2.0
    # via
    #   aiohttp
    #   eth-brownie
asttokens==2.0.5
    # via
    #   eth-brownie
    #   vyper
async-timeout==4.0.2
    # via
    #   aiohttp
    #   eth-brownie
attrs==21.4.0
    # via
    #   aiohttp
    #   eth-brownie
    #   hypothesis
    #   jsonschema
    #   pytest
base58==2.1.1
    # via
    #   eth-brownie
    #   multiaddr
bitarray==1.2.2
    # via
    #   eth-account
    #   eth-brownie
black==22.3.0
    # via
    #   -r requirements.in
    #   eth-brownie
brownie-token-tester==0.3.2
    # via -r requirements.in
certifi==2021.10.8
    # via
    #   eth-brownie
    #   requests
cfgv==3.3.1
    # via pre-commit
charset-normalizer==2.0.12
    # via
    #   aiohttp
    #   eth-brownie
    #   requests
click==8.1.3
    # via
    #   -r requirements.in
    #   black
    #   eth-brownie
    #   pip-tools
cytoolz==0.11.2
    # via
    #   eth-brownie
    #   eth-keyfile
    #   eth-utils
dataclassy==0.11.1
    # via
    #   eip712
    #   eth-brownie
distlib==0.3.4
    # via virtualenv
eip712==0.1.0
    # via eth-brownie
eth-abi==2.1.1
    # via
    #   eip712
    #   eth-account
    #   eth-brownie
    #   eth-event
    #   web3
eth-account==0.5.7
    # via
    #   eth-brownie
    #   web3
eth-brownie==1.18.2
    # via
    #   -r requirements.in
    #   brownie-token-tester
eth-event==1.2.3
    # via eth-brownie
eth-hash[pycryptodome]==0.3.2
    # via
    #   eth-brownie
    #   eth-event
    #   eth-utils
    #   web3
eth-keyfile==0.5.1
    # via
    #   eth-account
    #   eth-brownie
eth-keys==0.3.4
    # via
    #   eth-account
    #   eth-brownie
    #   eth-keyfile
eth-rlp==0.2.1
    # via
    #   eth-account
    #   eth-brownie
eth-typing==2.3.0
    # via
    #   eip712
    #   eth-abi
    #   eth-brownie
    #   eth-keys
    #   eth-utils
    #   web3
eth-utils==1.10.0
    # via
    #   eip712
    #   eth-abi
    #   eth-account
    #   eth-brownie
    #   eth-event
    #   eth-keyfile
    #   eth-keys
    #   eth-rlp
    #   rlp
    #   web3
execnet==1.9.0
    # via
    #   eth-brownie
    #   pytest-xdist
filelock==3.7.0
    # via virtualenv
flake8==3.9.2
    # via -r requirements.in
frozenlist==1.3.0
    # via
    #   aiohttp
    #   aiosignal
    #   eth-brownie
hexbytes==0.2.2
    # via
    #   eip712
    #   eth-account
    #   eth-brownie
    #   eth-event
    #   eth-rlp
    #   web3
hypothesis==6.27.3
    # via eth-brownie
identify==2.5.1
    # via pre-commit
idna==3.3
    # via
    #   eth-brownie
    #   requests
    #   yarl
inflection==0.5.0
    # via
    #   eth-brownie
    #   mythx-models
    #   pythx
iniconfig==1.1.1
    # via
    #   eth-brownie
    #   pytest
ipfshttpclient==0.8.0a2
    # via
    #   eth-brownie
    #   web3
isort==5.10.1
    # via -r requirements.in
jsonschema==3.2.0
    # via
    #   eth-brownie
    #   mythx-models
    #   web3
lazy-object-proxy==1.7.1
    # via eth-brownie
lru-dict==1.1.7
    # via
    #   eth-brownie
    #   web3
mccabe==0.6.1
    # via flake8
multiaddr==0.0.9
    # via
    #   eth-brownie
    #   ipfshttpclient
multidict==6.0.2
    # via
    #   aiohttp
    #   eth-brownie
    #   yarl
mypy-extensions==0.4.3
    # via
    #   black
    #   eth-brownie
mythx-models==1.9.1
    # via
    #   eth-brownie
    #   pythx
netaddr==0.8.0
    # via
    #   eth-brownie
    #   multiaddr
nodeenv==1.6.0
    # via pre-commit
packaging==21.3
    # via
    #   eth-brownie
    #   pytest
parsimonious==0.8.1
    # via
    #   eth-abi
    #   eth-brownie
pathspec==0.9.0
    # via
    #   black
    #   eth-brownie
pep517==0.12.0
    # via pip-tools
pip-tools==6.6.2
    # via -r requirements.in
platformdirs==2.5.2
    # via
    #   black
    #   eth-brownie
    #   virtualenv
pluggy==1.0.0
    # via
    #   eth-brownie
    #   pytest
pre-commit==2.19.0
    # via -r requirements.in
prompt-toolkit==3.0.29
    # via eth-brownie
protobuf==3.20.1
    # via
    #   eth-brownie
    #   web3
psutil==5.9.0
    # via eth-brownie
py==1.11.0
    # via
    #   eth-brownie
    #   pytest
    #   pytest-forked
py-solc-ast==1.2.9
    # via eth-brownie
py-solc-x==1.1.1
    # via eth-brownie
pycodestyle==2.7.0
    # via flake8
pycryptodome==3.14.1
    # via
    #   eip712
    #   eth-brownie
    #   eth-hash
    #   eth-keyfile
    #   vyper
pyflakes==2.3.1
    # via flake8
pygments==2.12.0
    # via
    #   eth-brownie
    #   pygments-lexer-solidity
pygments-lexer-solidity==0.7.0
    # via eth-brownie
pyjwt==1.7.1
    # via
    #   eth-brownie
    #   pythx
pyparsing==3.0.9
    # via
    #   eth-brownie
    #   packaging
pyrsistent==0.18.1
    # via
    #   eth-brownie
    #   jsonschema
pytest==6.2.5
    # via
    #   eth-brownie
    #   pytest-forked
    #   pytest-xdist
pytest-forked==1.4.0
    # via
    #   eth-brownie
    #   pytest-xdist
pytest-xdist==1.34.0
    # via eth-brownie
python-dateutil==2.8.1
    # via
    #   eth-brownie
    #   mythx-models
    #   pythx
python-dotenv==0.16.0
    # via eth-brownie
pythx==1.6.1
    # via eth-brownie
pyyaml==5.4.1
    # via
    #   eth-brownie
    #   pre-commit
requests==2.27.1
    # via
    #   -r requirements.in
    #   eth-brownie
    #   ipfshttpclient
    #   py-solc-x
    #   pythx
    #   vvm
    #   web3
rlp==2.0.1
    # via
    #   eth-account
    #   eth-brownie
    #   eth-rlp
semantic-version==2.8.5
    # via
    #   eth-brownie
    #   py-solc-x
    #   vvm
    #   vyper
six==1.16.0
    # via
    #   asttokens
    #   eth-brownie
    #   jsonschema
    #   multiaddr
    #   parsimonious
    #   pytest-xdist
    #   python-dateutil
    #   virtualenv
sortedcontainers==2.4.0
    # via
    #   eth-brownie
    #   hypothesis
toml==0.10.2
    # via
    #   eth-brownie
    #   pre-commit
    #   pytest
tomli==2.0.1
    # via
    #   black
    #   eth-brownie
    #   pep517
toolz==0.11.2
    # via
    #   cytoolz
    #   eth-brownie
tqdm==4.64.0
    # via eth-brownie
urllib3==1.26.9
    # via
    #   eth-brownie
    #   requests
varint==1.0.2
    # via
    #   eth-brownie
    #   multiaddr
virtualenv==20.14.1
    # via pre-commit
vvm==0.1.0
    # via eth-brownie
vyper==0.3.3
    # via eth-brownie
wcwidth==0.2.5
    # via
    #   eth-brownie
    #   prompt-toolkit
web3==5.29.0
    # via eth-brownie
websockets==9.1
    # via
    #   eth-brownie
    #   web3
wheel==0.37.1
    # via
    #   eth-brownie
    #   pip-tools
    #   vyper
wrapt==1.14.1
    # via eth-brownie
yarl==1.7.2
    # via
    #   aiohttp
    #   eth-brownie

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
